pyproject.toml
[project]
name = "ascend"
version = "0.1.0"
description = "Serverless cloud execution for Python"
authors = [{name = "Ascend Team"}]
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    # Core
    "cloudpickle>=3.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "packaging>=23.0",
    # Kubernetes
    "kubernetes>=29.0.0",
    # Cloud storage abstraction
    "fsspec>=2024.2",
    # Utilities
    "rich>=13.7.0",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.15.0",
    "azure-storage-blob>=12.19.0",
    "azure-mgmt-containerservice>=28.0.0",
    "azure-mgmt-storage>=21.1.0",
    "azure-mgmt-containerregistry>=10.3.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-containerregistry>=1.2.0",
    "azure-mgmt-msi>=7.0.0",
    "azure-mgmt-authorization>=4.0.0",
    "adlfs>=2024.4",
]
# Future:
# gcp = ["gcsfs>=2024.2", "google-cloud-container>=2.40"]
# aws = ["s3fs>=2024.2", "boto3>=1.34"]

[project.scripts]
ascend = "ascend.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ascend"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-timeout>=2.2.0",
    "python-dotenv>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "hypothesis>=6.92.0",
    "optuna>=4.7.0",
    "torch>=2.0.0",
    "pytorch-lightning>=2.0",
    "torchvision>=0.15.0",
    "torchmetrics>=1.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
markers = [
    "integration: Tests requiring real Azure/AKS infrastructure",
    "kind: Tests requiring local Kind cluster",
    "slow: Tests that take more than 30 seconds",
]
testpaths = ["tests"]
timeout = 300
